+++ models/models.py
@@ -27,7 +27,7 @@ class OccupancyEncoderCNN(nn.Module):
         for name in names[-6:]:
             setattr(self.fcn8, name, modified_last_layers[i])
             i += 1
-        self.fcn8 = self.fcn8.double()
+        self.fcn8 = self.fcn8.float()
         self.sigmoid = nn.Sigmoid()
 
     def forward(self, x):
@@ -58,7 +58,7 @@ class DetectionEncoderCNN(nn.Module):
             setattr(self.fcn8, name, modified_last_layers[i])
             i += 1
 
-        self.fcn8 = self.fcn8.double()
+        self.fcn8 = self.fcn8.float()
         self.sigmoid = nn.Sigmoid()
         
     def forward(self, x):
